requirements.txt
langchain
python-dotenv
streamlit
langchain_openai
langchain_community